sprint_4/requirements-lambda.txt
psycopg2-binary==2.9.9
jmespath==1.0.1
python-dotenv==1.0.1